Adjusting Movevment
Worked on making the movement feel better for the crab, also changed the button scheme you need to press for jumping off a wall. You are stuck on the wall until you jump, either you face the wall or away from it and jump then by pressing the movement key + jump key.